+++ src/components/sidebar/routePathView/routePathListTab/routePathListItem.scss
